+++ mod.ts
@@ -1,4 +1,4 @@
 export { Cookie } from "./cookie.ts";
 export type { CookieOptions } from "./cookie.ts";
 export { CookieJar } from "./cookie_jar.ts";
-export { wrapFetch } from "./fetch_wrapper.ts";
+export * from "./fetch_wrapper.ts";